requirements.txt
# Web Framework
fastapi[all]
uvicorn
gunicorn

# Database & Models
sqlmodel
psycopg2-binary
alembic

# Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Data Processing
pandas
numpy

# Configuration
python-dotenv
pydantic-settings

# Development & Testing
pytest
httpx
black
isort
flake8
